Remove DRI support to fix build after devPrivates rework
This driver uses devprivates of all kinds, but this is
only done in deprecated and unused code so we can simply
remove it all.

DRM/DRI support has been commented out for years, and was
done during the conversion over to XAA acceleration.  This
code would need to be essentially rewritten to work again,
so we can just remove this stuff for now.

The rest were either:

1) DRI/DRM related uses
2) the private allocation code
3) cases that could index to the pScrn to get the FFB private

And that's all fixed up here.

Signed-off-by: David S. Miller <davem@davemloft.net>